[FIX] CL warnings.
git-svn-id: https://b-tk.googlecode.com/svn/BTK/trunk@683 93be6e1e-6fcc-11de-baec-69240fdc299a

+++ Code/VisSupport/VTK/btkVTKSegmentsFramesSource.cpp
@@ -409,8 +409,8 @@ namespace btk
     int linkNumber = 0;
     for (std::list<SegmentDefinition>::const_iterator itS = this->m_Definitions.begin() ; itS != this->m_Definitions.end() ; ++itS)
     {
-      markerNumber += itS->markerIds.size();
-      linkNumber += itS->links.size();
+      markerNumber += static_cast<int>(itS->markerIds.size());
+      linkNumber += static_cast<int>(itS->links.size());
     }
     
     vtkIntArray* colors = vtkIntArray::New(); colors->SetName("Colors"); colors->SetNumberOfValues(linkNumber);
